Apply technical review from Hojin.Jang
Update boundary text part in HTTP/2 message example. For more detail, see below:

https://www.w3.org/Protocols/rfc1341/7_2_Multipart.html

Thus, a typical multipart Content-Type header field might look like this:

     Content-Type: multipart/mixed; 
          boundary=gc0p4Jq0M2Yt08jU534c0p
 
This indicates that the entity consists of several parts, each itself with a structure that is syntactically identical to an RFC 822 message, except that the header area might be completely empty, and that the parts are each preceded by the line
     --gc0p4Jq0M2Yt08jU534c0p

The encapsulation boundary following the last body part is a distinguished delimiter that indicates that no further body parts will follow. Such a delimiter is identical to the previous delimiters, with the addition of two more hyphens at the end of the line:

     --gc0p4Jq0M2Yt08jU534c0p--